requirements.lambda.txt
pandas==3.0.1
numpy==2.4.2
psycopg2-binary==2.9.11
requests==2.32.5
urllib3==2.6.3
python-dateutil==2.9.0.post0
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
six==1.17.0
tzdata==2025.3
xlrd
openpyxl
lxml